+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Partition Wizard is now able to convert to safeboot from Shelly partition layout (#19034)
 - Matter mini-profiler (#19075)
+- Berry `_class` can be used in `static var` initialization code
 
 ### Breaking Changed